Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a,b,c;
-    scanf("%f %f %f",&a,&b,&c);
+    int a,b,c;
+    scanf("%d %d %d",&a,&b,&c);
     (a>b && a>c)?printf("%d",a):(b>a && b>c)?printf("%d",b):printf("%d",c);
     return 0;
 }
